// src/models/post-links201-response.ts
export interface PostLinks201Response {
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    uuid: string;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    ownerId: string;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    alias: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    name: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    description: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    status: PostLinks201ResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    type: PostLinks201ResponseTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    expirationTime: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof PostLinks201Response
     */
    deleted: boolean;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    createdAt: string;
    /**
     * 
     * @type {string}
     * @memberof PostLinks201Response
     */
    updatedAt: string | null;
    /**
     * 
     * @type {GetLinksByLinkUuidTollsActive200Response}
     * @memberof PostLinks201Response
     */
    toll: GetLinksByLinkUuidTollsActive200Response;
    /**
     * 
     * @type {GetLinksByLinkUuidResourcesActive200Response}
     * @memberof PostLinks201Response
     */
    resource: GetLinksByLinkUuidResourcesActive200Response;
    /**
     * 
     * @type {Array<PostLinksByLinkUuidActions201Response>}
     * @memberof PostLinks201Response
     */
    actions: Array<PostLinksByLinkUuidActions201Response>;
}


/**
 * @export
 */
export const PostLinks201ResponseStatusEnum = {
    Active: 'active',
    Inactive: 'inactive',
    Used: 'used',
    Expired: 'expired'
} as const;
export type PostLinks201ResponseStatusEnum = typeof PostLinks201ResponseStatusEnum[keyof typeof PostLinks201ResponseStatusEnum];

/**
 * @export
 */
export const PostLinks201ResponseTypeEnum = {
    Onetime: 'onetime',
    Permanent: 'permanent'
} as const;
export type PostLinks201ResponseTypeEnum = typeof PostLinks201ResponseTypeEnum[keyof typeof PostLinks201ResponseTypeEnum];