// src/models/post-links-by-link-uuid-actions-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gatepay API
 * API documentation for Gatepay
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostLinksByLinkUuidActionsRequest
 */
export interface PostLinksByLinkUuidActionsRequest {
    /**
     * 
     * @type {string}
     * @memberof PostLinksByLinkUuidActionsRequest
     */
    type: PostLinksByLinkUuidActionsRequestTypeEnum;
    /**
     * 
     * @type {any}
     * @memberof PostLinksByLinkUuidActionsRequest
     */
    data?: any | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinksByLinkUuidActionsRequest
     */
    trigger: PostLinksByLinkUuidActionsRequestTriggerEnum;
}


/**
 * @export
 */
export const PostLinksByLinkUuidActionsRequestTypeEnum = {
    Callback: 'callback',
    Notification: 'notification',
    Email: 'email',
    Sms: 'sms'
} as const;
export type PostLinksByLinkUuidActionsRequestTypeEnum = typeof PostLinksByLinkUuidActionsRequestTypeEnum[keyof typeof PostLinksByLinkUuidActionsRequestTypeEnum];

/**
 * @export
 */
export const PostLinksByLinkUuidActionsRequestTriggerEnum = {
    PaymentSuccess: 'payment_success'
} as const;
export type PostLinksByLinkUuidActionsRequestTriggerEnum = typeof PostLinksByLinkUuidActionsRequestTriggerEnum[keyof typeof PostLinksByLinkUuidActionsRequestTriggerEnum];


/**
 * Check if a given object implements the PostLinksByLinkUuidActionsRequest interface.
 */
export function instanceOfPostLinksByLinkUuidActionsRequest(value: object): value is PostLinksByLinkUuidActionsRequest {
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('trigger' in value) || value['trigger'] === undefined) return false;
    return true;
}

export function PostLinksByLinkUuidActionsRequestFromJSON(json: any): PostLinksByLinkUuidActionsRequest {
    return PostLinksByLinkUuidActionsRequestFromJSONTyped(json, false);
}

export function PostLinksByLinkUuidActionsRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostLinksByLinkUuidActionsRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'type': json['type'],
        'data': json['data'] == null ? undefined : json['data'],
        'trigger': json['trigger'],
    };
}

export function PostLinksByLinkUuidActionsRequestToJSON(json: any): PostLinksByLinkUuidActionsRequest {
    return PostLinksByLinkUuidActionsRequestToJSONTyped(json, false);
}

export function PostLinksByLinkUuidActionsRequestToJSONTyped(value?: PostLinksByLinkUuidActionsRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'type': value['type'],
        'data': value['data'],
        'trigger': value['trigger'],
    };
}