// src/models/post-links-by-link-uuid-resources-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gatepay API
 * API documentation for Gatepay
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostLinksByLinkUuidResourcesRequest
 */
export interface PostLinksByLinkUuidResourcesRequest {
    /**
     * 
     * @type {string}
     * @memberof PostLinksByLinkUuidResourcesRequest
     */
    type: PostLinksByLinkUuidResourcesRequestTypeEnum;
    /**
     * 
     * @type {any}
     * @memberof PostLinksByLinkUuidResourcesRequest
     */
    data?: any | null;
}


/**
 * @export
 */
export const PostLinksByLinkUuidResourcesRequestTypeEnum = {
    Proxy: 'proxy',
    Link: 'link',
    Redirect: 'redirect',
    Tunnel: 'tunnel',
    Html: 'html'
} as const;
export type PostLinksByLinkUuidResourcesRequestTypeEnum = typeof PostLinksByLinkUuidResourcesRequestTypeEnum[keyof typeof PostLinksByLinkUuidResourcesRequestTypeEnum];


/**
 * Check if a given object implements the PostLinksByLinkUuidResourcesRequest interface.
 */
export function instanceOfPostLinksByLinkUuidResourcesRequest(value: object): value is PostLinksByLinkUuidResourcesRequest {
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function PostLinksByLinkUuidResourcesRequestFromJSON(json: any): PostLinksByLinkUuidResourcesRequest {
    return PostLinksByLinkUuidResourcesRequestFromJSONTyped(json, false);
}

export function PostLinksByLinkUuidResourcesRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostLinksByLinkUuidResourcesRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'type': json['type'],
        'data': json['data'] == null ? undefined : json['data'],
    };
}

export function PostLinksByLinkUuidResourcesRequestToJSON(json: any): PostLinksByLinkUuidResourcesRequest {
    return PostLinksByLinkUuidResourcesRequestToJSONTyped(json, false);
}

export function PostLinksByLinkUuidResourcesRequestToJSONTyped(value?: PostLinksByLinkUuidResourcesRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'type': value['type'],
        'data': value['data'],
    };
}