// src/models/post-links-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gatepay API
 * API documentation for Gatepay
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostLinksRequest
 */
export interface PostLinksRequest {
    /**
     * 
     * @type {string}
     * @memberof PostLinksRequest
     */
    alias?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinksRequest
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinksRequest
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinksRequest
     */
    expirationTime?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostLinksRequest
     */
    status?: PostLinksRequestStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof PostLinksRequest
     */
    type?: PostLinksRequestTypeEnum;
}


/**
 * @export
 */
export const PostLinksRequestStatusEnum = {
    Active: 'active',
    Inactive: 'inactive',
    Used: 'used',
    Expired: 'expired'
} as const;
export type PostLinksRequestStatusEnum = typeof PostLinksRequestStatusEnum[keyof typeof PostLinksRequestStatusEnum];

/**
 * @export
 */
export const PostLinksRequestTypeEnum = {
    Onetime: 'onetime',
    Permanent: 'permanent'
} as const;
export type PostLinksRequestTypeEnum = typeof PostLinksRequestTypeEnum[keyof typeof PostLinksRequestTypeEnum];


/**
 * Check if a given object implements the PostLinksRequest interface.
 */
export function instanceOfPostLinksRequest(value: object): value is PostLinksRequest {
    return true;
}

export function PostLinksRequestFromJSON(json: any): PostLinksRequest {
    return PostLinksRequestFromJSONTyped(json, false);
}

export function PostLinksRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostLinksRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'alias': json['alias'] == null ? undefined : json['alias'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'expirationTime': json['expirationTime'] == null ? undefined : json['expirationTime'],
        'status': json['status'] == null ? undefined : json['status'],
        'type': json['type'] == null ? undefined : json['type'],
    };
}

export function PostLinksRequestToJSON(json: any): PostLinksRequest {
    return PostLinksRequestToJSONTyped(json, false);
}

export function PostLinksRequestToJSONTyped(value?: PostLinksRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'alias': value['alias'],
        'name': value['name'],
        'description': value['description'],
        'expirationTime': value['expirationTime'],
        'status': value['status'],
        'type': value['type'],
    };
}